Updated PoW Tests
